Created new scripts.js with the JQuery code
If applied, this commit will have added the page scripts.js written in JQuery code to create a to do ist.
Was given a variable for the input.value and afterwards the inputValue variable was append to li variable.
Was created a condition to append any value insert by the user.
Was created a eventListener with double click the 'switch' on the class 'strike' that contains a line-through to mark the list element as 'done'.
Was created a crossOutButton to be activated through an eventListener on click that 'switch' on the 'delete' class and display:none the input.
Was created a sortable() function to apply in the list giving to the user the possibility to drag and move the 'li' elements.